+++ R/remotes.R
@@ -16,7 +16,7 @@ renv_remotes_resolve <- function(entry) {
 
   # check for URLs
   if (grepl("^(?:file|https?)://", entry))
-    return(renv_remotes_resolve_url(entry))
+    return(renv_remotes_resolve_url(entry, quiet = TRUE))
 
   # check for paths to existing local files
   local <-
@@ -323,7 +323,7 @@ renv_remotes_resolve_gitlab <- function(entry) {
 
 }
 
-renv_remotes_resolve_url <- function(entry) {
+renv_remotes_resolve_url <- function(entry, quiet = FALSE) {
 
   tempfile <- renv_tempfile("renv-url-")
   writeLines(entry, con = tempfile)
@@ -334,7 +334,7 @@ renv_remotes_resolve_url <- function(entry) {
   path <- renv_paths_source("url", name)
 
   ensure_parent_directory(path)
-  download(entry, path, quiet = TRUE)
+  download(entry, path, quiet = quiet)
 
   desc <- renv_description_read(path)