+++ NEWS.md
@@ -1,6 +1,8 @@
 
 # renv 0.9.0 (UNRELEASED)
 
+* `renv` is now able to restore packages with a recorded URL remote. (#272)
+
 * `renv::dependencies()` can now parse R package dependencies used as custom
   site generator in an Rmd yaml header. (#269, @cderv)